source-code/requirements.txt
smolagents
litellm[proxy]
accelerate
requests
beautifulsoup4
ollama
langchain
langchain-community
langchain-ollama
langgraph
rdflib
kuzu
langchain_openai
tabulate
kuzu
autogen-agentchat
autogen-ext[ollama]

playwright
browser_use

langgraph
langchain_community
duckduckgo-search